backend/requirements.txt
flask
flask-cors
flask
mysql-connector-python
watchdog
